sy_feature UserController 인터페이스 분리, SwaggerConfig 그룹 지정 해제
